08_generative_models/assignments/gan/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
pytorch-fid>=0.3.0
